Restructure folders, components and stories
